+++ src/components/presentational/prefCheckbox/PrefCheckbox.tsx
@@ -1,0 +1,26 @@
+import "./index.scss"
+import { FC } from "react"
+import { PrefCode, PrefName } from "@/constants/pref"
+
+type Props = {
+  onChange: (prefCode: PrefCode) => void
+  checkedIdList: PrefCode[]
+  prefName: PrefName
+  prefCode: PrefCode
+}
+
+export const PrefCheckbox: FC<Props> = ({ onChange, checkedIdList, prefName, prefCode }) => {
+  const checked = checkedIdList.includes(prefCode)
+  return (
+    <label htmlFor="pref_check_box" className="pref_checkbox--label">
+      <input
+        id="pref_check_box"
+        type="checkbox"
+        onChange={() => onChange(prefCode)}
+        checked={checked}
+        className="pref_checkbox--input"
+      />
+      <p className="pref_checkbox--text">{prefName}</p>
+    </label>
+  )
+}
